GDA-5883 ScanDataPoint scan number is now an int

+++ plugins/org.gda.tango/src/gda/device/detector/maxipix2/MaxiPix2MultiFrameDetector.java
@@ -454,7 +454,7 @@ public class MaxiPix2MultiFrameDetector extends DetectorBase implements Position
 		firstFrameOfScan=true;
 		ScanInformation currentScanInformation = InterfaceProvider.getCurrentScanInformationHolder().getCurrentScanInformation();
 		if( (currentScanInformation!=null) && useScanSpecificFolder  ){
-			Long scanNumber = currentScanInformation.getScanNumber();
+			int scanNumber = currentScanInformation.getScanNumber();
 			String dataDir = PathConstructor.createFromDefaultProperty();
 			dataDir = dataDir + File.separator + Long.toString(scanNumber) + File.separator + "mpx" + File.separator ;
 			File f = new File(dataDir);
